+++ libft/src/str/ft_strcpy.c
@@ -6,16 +6,16 @@
 /*   By: adelille <adelille@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/26 17:49:57 by adelille          #+#    #+#             */
-/*   Updated: 2021/09/27 15:22:40 by adelille         ###   ########.fr       */
+/*   Updated: 2022/06/09 11:10:39 by adelille         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../inc/libft.h"
 
-int	ft_strcpy(char *dst, const char *src)
+size_t	ft_strcpy(char *dst, const char *src)
 {
-	int	i;
-	int	size;
+	size_t	i;
+	size_t	size;
 
 	i = 0;
 	size = ft_strlen(src);
